media/videos/tars-ai-ethics/requirements.txt
# TARS Video Generator Requirements
# Python 3.8+

# Core video generation
moviepy==1.0.3
pillow>=10.0.0
numpy>=1.24.0

# Audio generation (multiple options)
gTTS>=2.3.0  # Free text-to-speech
pydub>=0.25.1  # Audio processing

# YouTube upload
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.80.0

# Utilities
requests>=2.28.0
python-dotenv>=1.0.0
tqdm>=4.65.0
